requirements.txt
certifi==2020.6.20
cycler==0.10.0
kiwisolver==1.2.0
llvmlite==0.34.0
matplotlib==3.3.2
numba==0.51.2
numpy==1.18.2
Pillow==8.0.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
six==1.14.0
